Use shared attribute types in generator
