+++ Praefixum.Demo/Praefixum.Demo.csproj
@@ -12,5 +12,6 @@
     <ProjectReference Include="..\Praefixum\Praefixum.csproj"
                       OutputItemType="Analyzer"
                       ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\Praefixum.Attributes\Praefixum.Attributes.csproj" />
   </ItemGroup>
 </Project>
