+++ Praefixum/UniqueIdGenerator.cs
@@ -9,50 +9,14 @@ using System.Text;
 
 namespace Praefixum;
 
-// Enum used by the source generator internally
-internal enum UniqueIdFormat
-{
-    Guid,
-    HtmlId,
-    Timestamp,
-    ShortHash
-}
+// Uses UniqueIdAttribute and UniqueIdFormat from Praefixum.Attributes
 
 [Generator]
 public sealed class PraefixumSourceGenerator : IIncrementalGenerator
 {
     public void Initialize(IncrementalGeneratorInitializationContext context)
-    {        // Register the attribute source first
-        context.RegisterPostInitializationOutput(ctx =>
-        {            ctx.AddSource("UniqueIdAttribute.g.cs", SourceText.From("""
-                #nullable enable
-                namespace Praefixum
-                {
-                    [System.AttributeUsage(System.AttributeTargets.Parameter)]
-                    public sealed class UniqueIdAttribute : System.Attribute
-                    {
-                        public UniqueIdFormat Format { get; }
-                        public string? Prefix { get; }
-                        public bool Deterministic { get; }
-
-                        public UniqueIdAttribute(UniqueIdFormat format = UniqueIdFormat.Guid, string? prefix = null, bool deterministic = true)
-                        {
-                            Format = format;
-                            Prefix = prefix;
-                            Deterministic = deterministic;
-                        }
-                    }
-
-                    public enum UniqueIdFormat
-                    {
-                        Guid,
-                        HtmlId,
-                        Timestamp,
-                        ShortHash
-                    }
-                }
-                """, Encoding.UTF8));
-        });
+    {
+        // Attribute definitions are provided by the Praefixum.Attributes assembly
 
         // Find method calls that need interception
         var methodCallsProvider = context.SyntaxProvider
